small refactoring of the files structure
