+++ src/components/news/GeneralNewsBlock.vue
